% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_neuroc_release.R
\name{current_neuroc_release}
\alias{current_neuroc_release}
\title{Current Neuroconductor Release Location}
\usage{
current_neuroc_release(secure = TRUE)
}
\arguments{
\item{secure}{Should https vs. http be used}
}
\value{
URL fo release page
}
\description{
Current Neuroconductor Release Location
}
\examples{
current_neuroc_release()
}
