% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_neuroc_release.R
\name{latest_neuroc_release}
\alias{latest_neuroc_release}
\title{Latest Neuroconductor Release Location}
\usage{
latest_neuroc_release(release = c("stable", "current"), secure = TRUE)
}
\arguments{
\item{release}{Stable or current (development) versions}

\item{secure}{Should https vs. http be used}
}
\value{
URL fo release page
}
\description{
Latest Neuroconductor Release Location
}
\examples{
latest_neuroc_release()
}
