% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_full_version.R
\name{make_full_version}
\alias{make_full_version}
\title{Make Full Package Version}
\usage{
make_full_version(x)
}
\arguments{
\item{x}{Character vector of package versions}
}
\value{
Character vector of versions, each with the same length.
}
\description{
Makes a package version to have all the same length.
This is helpful when using \code{\link{compareVersion}}.
}
\examples{
x = c("1.6", "1.6.0")
compareVersion(x[1], x[2])
x2 = make_full_version(x)
compareVersion(x2[1], x2[2])
x = c("1.6", "1.6.0")
compareVersion(x2[1], x2[2])
}
