% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_install.R
\name{neuro_install}
\alias{neuro_install}
\alias{neuroc_install}
\alias{neurocLite}
\title{Neuroconductor Installer}
\usage{
neuro_install(repo, release = c("stable", "current"),
  release_repo = latest_neuroc_release(), upgrade_dependencies = FALSE, ...)

neuroc_install(...)

neurocLite(...)
}
\arguments{
\item{repo}{Package name in neuroconductor}

\item{release}{Stable, development, or release version}

\item{release_repo}{Repo for release repository, passed to
\code{\link{install.packages}}.  If \code{release_repo = "github"},
then it will install using GitHub}

\item{upgrade_dependencies}{Should dependencies be updated?
passed to \code{\link[devtools]{install}} if using
\code{release_repo = "github"}}

\item{...}{additional arguments passed to
\code{\link[devtools]{install_github}} or
\code{\link{install.packages}} (for release)}
}
\value{
Result from \code{\link[devtools]{install_github}}
}
\description{
Install function for neuroconductor packages
}
