% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_package_table.R
\name{neuro_package_table}
\alias{neuro_package_table}
\title{Neuroconductor Package Table}
\usage{
neuro_package_table()
}
\value{
\code{data.frame} of packages with commit IDs
}
\description{
Returns the table of Neuroconductor packages
}
\examples{
neuro_package_table()
}

