% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_package_table.R
\name{neuro_packages}
\alias{neuro_packages}
\title{Neuroconductor Packages}
\usage{
neuro_packages(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{neuro_package_table}}}
}
\value{
\code{vector} of packages available on Neuroconductor
}
\description{
Returns the vector of Neuroconductor packages
}
\examples{
neuro_packages()
}
