% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_install.R
\name{neuro_install}
\alias{neuro_install}
\alias{neurocLite}
\alias{neuroc_install}
\title{Neuroconductor Installer}
\usage{
neuro_install(repo, release = c("stable", "development"), ...)

neuroc_install(...)

neurocLite(...)
}
\arguments{
\item{repo}{Package name in neuroconductor}

\item{release}{Stable or development version}

\item{...}{additional arguments passed to
\code{\link[devtools]{install_github}}}
}
\value{
Result from \code{\link[devtools]{install_github}}
}
\description{
Install function for neuroconductor packages
}

