% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_install.R
\name{neuro_install}
\alias{neuro_install}
\alias{neuroc_install}
\alias{neurocLite}
\title{Neuroconductor Installer}
\usage{
neuro_install(repo, release = c("release", "stable", "current"),
  release_repo = current_neuroc_release(), upgrade_dependencies = FALSE,
  ...)

neuroc_install(...)

neurocLite(...)
}
\arguments{
\item{repo}{Package name in neuroconductor}

\item{release}{Release, stable or development version}

\item{release_repo}{Repo for release repository, passed to
\code{\link{install.packages}}}

\item{upgrade_dependencies}{Should dependencies be updated?
passed to \code{\link[devtools]{install}}}

\item{...}{additional arguments passed to
\code{\link[devtools]{install_github}} or
\code{\link{install.packages}} (for release)}
}
\value{
Result from \code{\link[devtools]{install_github}}
}
\description{
Install function for neuroconductor packages
}
