% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_install.R
\name{neuro_install}
\alias{neuro_install}
\alias{neuroc_install}
\alias{neurocLite}
\title{Neuroconductor Installer}
\usage{
neuro_install(repo, release = c("stable", "current"),
  upgrade_dependencies = FALSE, ...)

neuroc_install(...)

neurocLite(...)
}
\arguments{
\item{repo}{Package name in neuroconductor}

\item{release}{Stable or development version}

\item{upgrade_dependencies}{Should dependencies be updated?
passed to \code{\link[devtools]{install}}}

\item{...}{additional arguments passed to
\code{\link[devtools]{install_github}}}
}
\value{
Result from \code{\link[devtools]{install_github}}
}
\description{
Install function for neuroconductor packages
}
