% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_package_table.R
\name{neuro_package_table}
\alias{neuro_package_table}
\title{Neuroconductor Package Table}
\usage{
neuro_package_table(
  path = "https://neuroconductor.org/neurocPackages",
  long = FALSE,
  deployment = FALSE
)
}
\arguments{
\item{path}{Path to the table of package}

\item{long}{Should the data be "long" (with respect to stable/current)}

\item{deployment}{indicator if this is a release, not standard flag.}
}
\value{
\code{data.frame} of packages with commit IDs
}
\description{
Returns the table of Neuroconductor packages
}
\note{
Package information is obtained from
\url{https://neuroconductor.org/neurocPackages}
}
\examples{
neuro_package_table()
}
